pyproject.toml
[tool.poetry]
name = "lang-chain-e2e"
version = "0.1.0"
description = ""
authors = ["Ryan McGuinness <rmcguinness@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


pypdf = "^4.3.1"
chromadb = "^0.5.5"
langchain-ollama = "^0.2.0"
langchain-community = "^0.3.2"
sentence-transformers = "^3.2.1"
langchain-huggingface = "^0.1.0"
langchain-chroma = "^0.1.4"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
example = 'lang_chain_e2e.cli:main'
